make it better

+++ db/migrate/20150918142545_add_total_hrs_to_task_logs.rb
@@ -1,5 +1,5 @@
 class AddTotalHrsToTaskLogs < ActiveRecord::Migration
   def change
-    add_column :task_logs, :total_hrs, :integer
+    add_column :task_logs, :total_hrs, :float
   end
 end
